move to component vs feature
