+++ src/components/Lightbox/chrome/Footer.tsx
