+++ src/components/Lightbox/chrome/Header.tsx
