+++ src/components/Lightbox/chrome/ImageMenu.tsx
