+++ src/components/Lightbox/chrome/PagerDots.tsx
