+++ src/components/Lightbox/index.tsx
