+++ src/components/Lightbox/pager/ImageItem/ImageItem.android.tsx
