+++ src/components/Lightbox/pager/ImageItem/ImageItem.ios.tsx
