+++ src/components/Lightbox/pager/ImageItem/ImageItem.tsx
