+++ src/components/Lightbox/pager/ImagePager.tsx
@@ -36,8 +36,8 @@ import * as ScreenOrientation from 'expo-screen-orientation'
 import {type Dimensions} from '#/lib/media/types'
 import {useTheme} from '#/alf'
 import {setSystemUITheme} from '#/alf/util/systemUI'
+import {type Lightbox} from '#/components/Lightbox/state'
 import {IS_IOS} from '#/env'
-import {type Lightbox} from '#/features/lightbox/state'
 import {PlatformInfo} from '../../../../modules/expo-bluesky-swiss-army'
 import {Footer} from '../chrome/Footer'
 import {Header} from '../chrome/Header'
