+++ src/components/Lightbox/pager/transforms.ts
