+++ src/components/Lightbox/types.ts
