+++ src/components/Post/Embed/ImageEmbed.tsx
@@ -6,10 +6,10 @@ import {atoms as a, tokens} from '#/alf'
 import {AutoSizedImage} from '#/components/images/AutoSizedImage'
 import {Gallery} from '#/components/images/Gallery'
 import {ImageLayoutGrid} from '#/components/images/ImageLayoutGrid'
+import {useLightboxControls} from '#/components/Lightbox/state'
+import {type Dimensions} from '#/components/Lightbox/types'
 import {PostEmbedViewContext} from '#/components/Post/Embed/types'
 import {useAnalytics} from '#/analytics'
-import {useLightboxControls} from '#/features/lightbox/state'
-import {type Dimensions} from '#/features/lightbox/types'
 import {type EmbedType} from '#/types/bsky/post'
 import {type CommonProps} from './types'
