+++ src/state/util.ts
@@ -1,7 +1,7 @@
 import {useCallback} from 'react'
 
 import {useDialogStateControlContext} from '#/state/dialogs'
-import {useLightboxControls} from '#/features/lightbox/state'
+import {useLightboxControls} from '#/components/Lightbox/state'
 import {useModalControls} from './modals'
 import {useComposerControls} from './shell/composer'
 import {useSetDrawerOpen} from './shell/drawer-open'
